Move do_at_generic_first_appt outside of core
