+++ src/tlo/methods/alri.py
@@ -36,6 +36,7 @@ from tlo.lm import LinearModel, LinearModelType, Predictor
 from tlo.methods import Metadata
 from tlo.methods.causes import Cause
 from tlo.methods.hsi_event import HSI_Event
+from tlo.methods.hsi_generic_first_appts import GenericFirstApptModule
 from tlo.methods.symptommanager import Symptom
 from tlo.util import random_date, sample_outcome
 
@@ -54,7 +55,7 @@ to_prob = lambda odds: odds / (1.0 + odds)  # noqa: E731
 # ---------------------------------------------------------------------------------------------------------
 
 
-class Alri(Module):
+class Alri(GenericFirstApptModule):
     """This is the disease module for Acute Lower Respiratory Infections."""
 
     INIT_DEPENDENCIES = {
