+++ src/tlo/methods/malaria.py
@@ -11,25 +11,26 @@ from typing import TYPE_CHECKING, List, Literal, Optional, Union
 
 import pandas as pd
 
-from tlo import DateOffset, Module, Parameter, Property, Types, logging
+from tlo import DateOffset, Parameter, Property, Types, logging
 from tlo.events import Event, IndividualScopeEventMixin, PopulationScopeEventMixin, RegularEvent
 from tlo.lm import LinearModel, Predictor
 from tlo.methods import Metadata
 from tlo.methods.causes import Cause
 from tlo.methods.dxmanager import DxTest
 from tlo.methods.hsi_event import HSI_Event
+from tlo.methods.hsi_generic_first_appts import GenericFirstApptModule
 from tlo.methods.symptommanager import Symptom
 from tlo.util import random_date
 
 if TYPE_CHECKING:
-    from tlo.core import DiagnosisFunction
+    from tlo.methods.hsi_generic_first_appts import DiagnosisFunction
     from tlo.population import IndividualProperties
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 
-class Malaria(Module):
+class Malaria(GenericFirstApptModule):
     def __init__(self, name=None, resourcefilepath=None):
         """Create instance of Malaria module
