+++ src/tlo/methods/measles.py
@@ -4,11 +4,12 @@ from typing import List
 
 import pandas as pd
 
-from tlo import DateOffset, Module, Parameter, Property, Types, logging
+from tlo import DateOffset, Parameter, Property, Types, logging
 from tlo.events import Event, IndividualScopeEventMixin, PopulationScopeEventMixin, RegularEvent
 from tlo.methods import Metadata
 from tlo.methods.causes import Cause
 from tlo.methods.hsi_event import HSI_Event
+from tlo.methods.hsi_generic_first_appts import GenericFirstApptModule
 from tlo.methods.symptommanager import Symptom
 from tlo.util import random_date
 
@@ -16,7 +17,7 @@ logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 
-class Measles(Module):
+class Measles(GenericFirstApptModule):
     """This module represents measles infections and disease."""
 
     INIT_DEPENDENCIES = {'Demography', 'HealthSystem', 'SymptomManager'}
