+++ src/tlo/methods/stunting.py
@@ -23,6 +23,7 @@ from tlo.events import IndividualScopeEventMixin, PopulationScopeEventMixin, Reg
 from tlo.lm import LinearModel, LinearModelType, Predictor
 from tlo.methods import Metadata
 from tlo.methods.hsi_event import HSI_Event
+from tlo.methods.hsi_generic_first_appts import GenericFirstApptModule
 
 if TYPE_CHECKING:
     from tlo.population import IndividualProperties
@@ -35,7 +36,7 @@ logger.setLevel(logging.INFO)
 #   MODULE DEFINITION
 # ---------------------------------------------------------------------------------------------------------
 
-class Stunting(Module):
+class Stunting(GenericFirstApptModule):
     """This is the disease module for Stunting"""
 
     INIT_DEPENDENCIES = {'Demography', 'Wasting', 'NewbornOutcomes', 'Diarrhoea', 'Hiv'}
